refactor(test): move tests to separate file
